remaster: full speed-run overhaul
Removed deprecated assets, refactored all components with useCallback/memo and //Note; comments, unified UI styling, consolidated Handicap feature with new logic, extended GraphQL schema & cleaned auth/utils/server setup.
