+++ server/src/utils/weather.ts
